Move JsonMapExtension into type_utils.dart
